# optimized normalize_min_max and denormalize