--- a/kornia/enhance/normalize.py
+++ b/kornia/enhance/normalize.py
@@ -258,13 +258,20 @@ def denormalize(data: Tensor, mean: Union[Tensor, float], std: Union[Tensor, flo
         mean = torch.as_tensor(mean, device=data.device, dtype=data.dtype)
         std = torch.as_tensor(std, device=data.device, dtype=data.dtype)
 
-    mean = mean[..., None]
-    std = std[..., None]
+    if mean.dim() == 1:
+        mean = mean.view(1, -1, *([1] * (data.dim() - 2)))
+    # If the tensor is >1D (e.g., (B, C)), reshape to (B, C, 1, ...)
+    else:
+        while len(mean.shape) < data.dim():
+            mean = mean.unsqueeze(-1)
+            
+    if std.dim() == 1:
+        std = std.view(1, -1, *([1] * (data.dim() - 2)))
+    else:
+        while len(std.shape) < data.dim():
+            std = std.unsqueeze(-1)
 
-    out: Tensor = (data.view(shape[0], shape[1], -1) * std) + mean
-
-    return out.view(shape)
-
+    return torch.addcmul(mean, data, std)
 
 def normalize_min_max(x: Tensor, min_val: float = 0.0, max_val: float = 1.0, eps: float = 1e-6) -> Tensor:
     r"""Normalise an image/video tensor by MinMax and re-scales the value between a range.
@@ -301,17 +308,19 @@ def normalize_min_max(x: Tensor, min_val: float = 0.0, max_val: float = 1.0, eps
         raise TypeError(f"'min_val' should be a float. Got: {type(min_val)}.")
 
     if not isinstance(max_val, float):
-        raise TypeError(f"'b' should be a float. Got: {type(max_val)}.")
+        raise TypeError(f"'max_val' should be a float. Got: {type(max_val)}.")
 
     if len(x.shape) < 3:
         raise ValueError(f"Input shape must be at least a 3d tensor. Got: {x.shape}.")
+    
+    B, C = x.shape[:2]
+    x_flat = x.view(B, C, -1)
 
-    shape = x.shape
-    B, C = shape[0], shape[1]
+    #TODO replace with aminmax when gradient is supported
+    x_min = x_flat.min(-1, keepdim=True)[0]
+    x_max = x_flat.max(-1, keepdim=True)[0]
 
-    x_reshaped = x.view(B, C, -1)
-    x_min = x_reshaped.min(-1, keepdim=True)[0]  # Shape: (B, C, 1)
-    x_max = x_reshaped.max(-1, keepdim=True)[0]  # Shape: (B, C, 1)
-
-    x_out = (max_val - min_val) * (x_reshaped - x_min) / (x_max - x_min + eps) + min_val
-    return x_out.view(shape)
+    scale = (max_val - min_val) / (x_max - x_min + eps)
+    bias = min_val - x_min * scale
+    x_flat = torch.addcmul(bias, x_flat, scale)    
+    return x_flat.view_as(x)