fuses the main calculation of normalize min_max and denormalize to use addcmul, giving a speedup of around 2x on average, theres one further optimization possible in normalize_min_max by using aminmax but gradients aren't supported for that rn, soo i added a todo there

Original function average time:   1.091018 ms
Fused function average time:      0.454880 ms

🚀 Speedup: 2.40x

Original function average time: 0.8768 ms
Fused function average time:    0.4209 ms

🚀 Speedup: 2.08x

https://colab.research.google.com/drive/1roNaIujGDWgabjRlfJtmCh-RbEbFylFn?usp=sharing

heres the benchmarking and validation script for the same